+++ bees.c
@@ -20,34 +20,34 @@ int getTrial(Bees bees, int i)
 	return bees->trial[i];
 }
 
-void setFitness(Bees bees, int i, float newFitness)
+void setFitness(Bees bees, int i, double newFitness)
 {
 	bees->fitness[i] = newFitness;
 }
 
-float getFitness(Bees bees, int i)
+double getFitness(Bees bees, int i)
 {
 	return bees->fitness[i];
 }
 
-void setP(Bees bees, int i, float newP)
+void setP(Bees bees, int i, double newP)
 {
 	bees->p[i] = newP;
 }
 
-float getP(Bees bees, int i)
+double getP(Bees bees, int i)
 {
 	return bees->p[i];
 }
 
-void setPosition(Bees bees, int i, float newPosition[])
+void setPosition(Bees bees, int i, double newPosition[])
 {
 	int y;
 	for (y=0; y<D; y++)
 		bees->positions[i][y] = newPosition[y];
 }
 
-float *getPosition(Bees bees, int i)
+double *getPosition(Bees bees, int i)
 {
 	return bees->positions[i];
 }
